Combine multi_rotation tests with rotation tests
